+++ examples/blog/jobPosting/src/pages/companies/create.tsx
@@ -10,8 +10,6 @@ import {
     Col,
 } from "@pankod/refine";
 
-import "react-mde/lib/styles/css/react-mde-all.css";
-
 import { ICompany } from "interfaces";
 
 export const CompanyCreate: React.FC<IResourceComponentsProps> = () => {