+++ examples/dataProvider/nestjsxCrud/src/pages/posts/edit.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import { IResourceComponentsProps, useApiUrl } from "@pankod/refine-core";
 
 import {
@@ -12,10 +12,7 @@ import {
     useFileUploadState,
 } from "@pankod/refine-antd";
 
-import ReactMarkdown from "react-markdown";
-import ReactMde from "react-mde";
-
-import "react-mde/lib/styles/css/react-mde-all.css";
+import MDEditor from "@uiw/react-md-editor";
 
 import { IPost, ICategory, ITags } from "interfaces";
 import { normalizeFile } from "utility/normalize";
@@ -36,10 +33,6 @@ export const PostEdit: React.FC<IResourceComponentsProps> = () => {
     });
 
     const { isLoading, onChange } = useFileUploadState();
-
-    const [selectedTab, setSelectedTab] = useState<"write" | "preview">(
-        "write",
-    );
 
     return (
         <Edit saveButtonProps={{ ...saveButtonProps, disabled: isLoading }}>
@@ -120,15 +113,7 @@ export const PostEdit: React.FC<IResourceComponentsProps> = () => {
                         },
                     ]}
                 >
-                    <ReactMde
-                        selectedTab={selectedTab}
-                        onTabChange={setSelectedTab}
-                        generateMarkdownPreview={(markdown) =>
-                            Promise.resolve(
-                                <ReactMarkdown>{markdown}</ReactMarkdown>,
-                            )
-                        }
-                    />
+                    <MDEditor data-color-mode="light" />
                 </Form.Item>
                 <Form.Item label="Images">
                     <Form.Item