+++ examples/tutorial/chakra-ui/src/components/pagination/index.tsx
@@ -1,0 +1,62 @@
+import { FC } from "react";
+import { HStack, Button, Box } from "@chakra-ui/react";
+import { IconChevronRight, IconChevronLeft } from "@tabler/icons";
+import { IconButton, usePagination } from "@pankod/refine-chakra-ui";
+
+type PaginationProps = {
+    current: number;
+    pageCount: number;
+    setCurrent: (page: number) => void;
+};
+
+export const Pagination: FC<PaginationProps> = ({
+    current,
+    pageCount,
+    setCurrent,
+}) => {
+    const pagination = usePagination({
+        current,
+        pageCount,
+    });
+
+    return (
+        <Box display="flex" justifyContent="flex-end">
+            <HStack my="3" spacing="1">
+                {pagination?.prev && (
+                    <IconButton
+                        aria-label="previous page"
+                        onClick={() => setCurrent(current - 1)}
+                        disabled={!pagination?.prev}
+                        variant="outline"
+                    >
+                        <IconChevronLeft size="18" />
+                    </IconButton>
+                )}
+
+                {pagination?.items.map((page) => {
+                    if (typeof page === "string")
+                        return <span key={page}>...</span>;
+
+                    return (
+                        <Button
+                            key={page}
+                            onClick={() => setCurrent(page)}
+                            variant={page === current ? "solid" : "outline"}
+                        >
+                            {page}
+                        </Button>
+                    );
+                })}
+                {pagination?.next && (
+                    <IconButton
+                        aria-label="next page"
+                        onClick={() => setCurrent(current + 1)}
+                        variant="outline"
+                    >
+                        <IconChevronRight size="18" />
+                    </IconButton>
+                )}
+            </HStack>
+        </Box>
+    );
+};