+++ packages/guesser/src/components/code-highlight/index.tsx
@@ -11,7 +11,11 @@ export const CodeHighlight: React.FC<{
 }> = ({ code, wrapperProps }) => {
     return (
         <div
-            style={{ maxHeight: "50vh", height: "100%", overflow: "scroll" }}
+            style={{
+                maxHeight: "50vh",
+                height: "100%",
+                overflow: "scroll",
+            }}
             {...(wrapperProps ?? {})}
         >
             <Highlight {...defaultProps} code={code} language="tsx">
@@ -24,7 +28,14 @@ export const CodeHighlight: React.FC<{
                 }) => (
                     <pre
                         className={className}
-                        style={{ ...style, padding: "14px", marginBottom: "0" }}
+                        style={{
+                            ...style,
+                            padding: "14px 14px 14px 14px",
+                            margin: "0",
+                            width: "100%",
+                            boxSizing: "border-box",
+                            overflowY: "scroll",
+                        }}
                     >
                         {tokens.map((line, i) => (
                             <div key={i} {...getLineProps({ line, key: i })}>