+++ packages/guesser/src/index.tsx
@@ -1,2 +1,3 @@
 export { createGuesser } from "./create-guesser";
 export { AntdShowGuesser, AntdEditGuesser, AntdListGuesser } from "@/guessers";
+export { MuiShowGuesser, MuiEditGuesser, MuiListGuesser } from "@/guessers";