+++ packages/mantine/src/contexts/form-context.ts
@@ -1,5 +1,5 @@
 import { createFormContext } from "@mantine/form";
 
-const [FormProvider, useFormContext, useForm] = createFormContext();
+const [FormProvider, useFormContext, useForm] = createFormContext<unknown>();
 
 export const FormContext = { FormProvider, useFormContext, useForm };