+++ packages/chakra-ui/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @pankod/refine-chakra-ui
 
+## 1.2.10
+
+### Patch Changes
+
+-   [#2969](https://github.com/refinedev/refine/pull/2969) [`a9459550a4`](https://github.com/refinedev/refine/commit/a9459550a49a640c5a1e393d4f2b8e6e9cd53dc6) Thanks [@omeraplak](https://github.com/omeraplak)! - Fixed peerDependencies of packages
+
+## 1.2.9
+
+### Patch Changes
+
+-   [#2969](https://github.com/refinedev/refine/pull/2969) [`a9459550a4`](https://github.com/refinedev/refine/commit/a9459550a49a640c5a1e393d4f2b8e6e9cd53dc6) Thanks [@omeraplak](https://github.com/omeraplak)! - Fixed peerDependencies of packages
+
+## 1.2.8
+
+### Patch Changes
+
+-   [#2970](https://github.com/refinedev/refine/pull/2970) [`513c078df1`](https://github.com/refinedev/refine/commit/513c078df1aa7b694fc41e5d710eff0d9a716fed) Thanks [@salihozdemir](https://github.com/salihozdemir)! - Fixed responsive style for error page.
+
+## 1.2.7
+
+### Patch Changes
+
+-   [#2970](https://github.com/refinedev/refine/pull/2970) [`513c078df1`](https://github.com/refinedev/refine/commit/513c078df1aa7b694fc41e5d710eff0d9a716fed) Thanks [@salihozdemir](https://github.com/salihozdemir)! - Fixed responsive style for error page.
+
+## 1.2.6
+
+### Patch Changes
+
+-   [#2959](https://github.com/refinedev/refine/pull/2959) [`3cd13fa5c2`](https://github.com/refinedev/refine/commit/3cd13fa5c24a38c3a14458e26a8a43567144b4ea) Thanks [@salihozdemir](https://github.com/salihozdemir)! - - Fixed error and ready pages issues that dark mode compatibility and mobile view issues.
+    -   Fixed crud components mobile view issues and spacing issues.
+    -   Update `Save` button icon button color and variant.
+    -   Rename `TextFieldComponent` export name to `TextField`.
+    -   Update `DeleteButton` color scheme.
+    -   Add missing props to `RefreshButton`'s `IconButton`.
+
+## 1.2.5
+
+### Patch Changes
+
+-   [#2959](https://github.com/refinedev/refine/pull/2959) [`3cd13fa5c2`](https://github.com/refinedev/refine/commit/3cd13fa5c24a38c3a14458e26a8a43567144b4ea) Thanks [@salihozdemir](https://github.com/salihozdemir)! - - Fixed error and ready pages issues that dark mode compatibility and mobile view issues.
+    -   Fixed crud components mobile view issues and spacing issues.
+    -   Update `Save` button icon button color and variant.
+    -   Rename `TextFieldComponent` export name to `TextField`.
+    -   Update `DeleteButton` color scheme.
+    -   Add missing props to `RefreshButton`'s `IconButton`.
+
+## 1.2.4
+
+### Patch Changes
+
+-   [#2948](https://github.com/refinedev/refine/pull/2948) [`add3da4c76`](https://github.com/refinedev/refine/commit/add3da4c76472d6c025aa3a2f0f21ecd3ab1589e) Thanks [@yildirayunlu](https://github.com/yildirayunlu)! - Fix AuthPage styles props and dark theme colors.
+
+## 1.2.3
+
+### Patch Changes
+
+-   [#2948](https://github.com/refinedev/refine/pull/2948) [`add3da4c76`](https://github.com/refinedev/refine/commit/add3da4c76472d6c025aa3a2f0f21ecd3ab1589e) Thanks [@yildirayunlu](https://github.com/yildirayunlu)! - Fix AuthPage styles props and dark theme colors.
+
 ## 1.2.2
 
 ### Patch Changes